pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cvr"
version = "0.1.0"
description = "Cockpit Voice Recorder - AI-powered video subtitle generator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "OpenAlchemy" }]
keywords = ["subtitle", "whisper", "translation", "srt", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
cvr = "cvr.cli:main"

[tool.ruff]
line-length = 100
